improved user following and extracted user profile to component

+++ api/user/src/main/java/com/foodagram/user/repository/UsersEngagementRepository.java
@@ -9,8 +9,8 @@ import java.util.UUID;
 
 @Repository
 public interface UsersEngagementRepository extends JpaRepository<UsersEngagement, UUID> {
-    long countUsersEngagementsByUserIdAndUserEngagementType(UUID userId, UserEngagementType userEngagementType);
-    long countUsersEngagementsByEngagedUserIdAndUserEngagementType(UUID engagedUserId, UserEngagementType userEngagementType);
+    int countUsersEngagementsByUserIdAndUserEngagementType(UUID userId, UserEngagementType userEngagementType);
+    int countUsersEngagementsByEngagedUserIdAndUserEngagementType(UUID engagedUserId, UserEngagementType userEngagementType);
 
     List<UsersEngagement> getUsersEngagementsByUserIdAndUserEngagementType(UUID userId, UserEngagementType userEngagementType);
     List<UsersEngagement> getUsersEngagementsByEngagedUserIdAndUserEngagementType(UUID engagedUserId, UserEngagementType userEngagementType);
